+++ server/item.hpp
@@ -66,7 +66,6 @@ public:
         database.execute(params,
                 "DELETE FROM `items` WHERE id = ? and userid = ?");
         uint64_t affected = database.affected_rows();
-        LOG_DEBUG("Affected: %d", affected);
         log_assert(affected <= 1);
         if (affected == 0)
             throw RowDoesNotExists("Can't destroy item, user does not "